Updated sln file and updated namespaces in files in the folder Pages Co-authored-by: Henriette <hoes@itu.dk> Co-authored-by: Laura <llsa@itu.dk>
