+++ src/UserFacade/Pages/Public.cshtml
@@ -1,13 +1,12 @@
-﻿@page "/{author}"
-@model Chirp.Pages.UserTimelineModel
+﻿@page "/"
+@model Chirp.src.Pages.PublicModel
 @{
     ViewData["Title"] = "Chirp!";
     Layout = "Shared/_Layout";
-    var routeName = HttpContext.GetRouteValue("author");
 }
 
 <div>
-    <h2> @routeName's Timeline </h2>
+    <h2> Public Timeline </h2>
 
     @if (Model.Cheeps.Any())
     {
