+++ src/UserFacade/Pages/UserTimeline.cshtml.cs
@@ -1,21 +1,21 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
-namespace Chirp.Pages;
+namespace Chirp.src.Pages;
 
-public class PublicModel : PageModel
+public class UserTimelineModel : PageModel
 {
     private readonly ICheepService _service;
     public List<CheepObject> Cheeps { get; set; }
 
-    public PublicModel(ICheepService service)
+    public UserTimelineModel(ICheepService service)
     {
         _service = service;
     }
 
-    public ActionResult OnGet()
+    public ActionResult OnGet(string author)
     {
-        Cheeps = _service.GetCheeps();
+        Cheeps = _service.GetCheepsFromAuthor(author);
         return Page();
     }
 }
